Include operational attributes in LDAP search results
This allows filtering users based on [operational attributes][1].

A good example of an operational attribute is the "memberOf" attribute in OpenLDAP setups.
While some LDAP servers may return this particular attribute as a normal, user attribute,
OpenLDAP configured with the memberof overlay considers it an operational attribute and
will return it only if that is asked explicitly.

Note: ldap3 library in version 0.9.9.2 introduced a bug when both `attributes=ALL_ATTRIBUTES` and
`get_operational_attributes=True` are used in search that makes search results include only the
operational attributes, so we exclude it from the requirements. This bug has been fixed for next
release.

[1]: https://tools.ietf.org/html/rfc4512#section-3.4